communications-service/requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
sqlalchemy>=2.0.34,<3.0.0
alembic>=1.13.2,<2.0.0
psycopg2-binary>=2.9.9
psycopg[binary]>=3.1.0
pyjwt>=2.8.0,<3.0.0
python-jose[cryptography]>=3.3.0
loguru>=0.7.2,<1.0.0
structlog>=23.2.0,<25.0.0
sendgrid>=6.11.0,<7.0.0
twilio>=8.10.3,<10.0.0
pika>=1.3.2,<2.0.0
python-multipart>=0.0.6,<1.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.2,<1.0.0
cryptography>=41.0.7
email-validator>=2.1.1,<3.0.0